Accounts.h

+++ Accounts.h
@@ -8,7 +8,7 @@ class account
     friend class bank;
   
     public:
-        account(string argf = "", string argl = "", int phone = 0, string email = "", string address = "", float balance = 0.0F);  // Default constructor -- Creates account
+        account(string argf = "", string argl = "", int long phone = 0, string email = "", string address = "", float balance = 0.0F);  // Default constructor -- Creates account
         ~account();                                         // Destructor
